# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ExecutionTarget": {
        "description": "Execution target.",
        "enum": [
          "pool",
          "on_demand"
        ],
        "title": "ExecutionTarget",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobKind": {
        "description": "Job kind.",
        "enum": [
          "replay",
          "session_run"
        ],
        "title": "JobKind",
        "type": "string"
      },
...
      "JobStatus": {
        "description": "Job status.",
        "enum": [
          "pending",
          "claimed",
          "running",
          "completed",
          "failed",
          "timed_out",
          "canceled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "Page_JobResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[JobResponse]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kitaru",
    "version": "0.21.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/experiment-runs/{run_id}/jobs": {
      "get": {
        "description": "List the jobs of an experiment run.\n\nClients observe HTTP 200 on success, 404 when no experiment run has\nthis id, and 422 on invalid filter or pagination parameters.\n\nArgs:\n    run_id: Id of the experiment run.\n    service: Experiment run service.\n    actor: Caller context.\n    status: Filter on job status.\n    page: Page number.\n    page_size: Page size.\n\nReturns:\n    Page of jobs.",
        "operationId": "list_experiment_run_jobs_v1_experiment_runs__run_id__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiment Run Jobs",
        "tags": [
          "experiment-runs"
        ]
      }
    },
...
    "/v1/jobs": {
      "get": {
        "description": "List jobs.\n\nClients observe HTTP 200 on success and 422 on invalid pagination\nparameters.\n\nArgs:\n    service: Job service.\n    actor: Caller context.\n    experiment_run_id: Filter on experiment run id.\n    original_session_id: Filter on the replayed session id.\n    kind: Filter on job kind.\n    job_status: Filter on job status.\n    standalone: Filter on standalone jobs.\n    worker_id: Filter on the claiming worker id.\n    execution_target: Filter on execution target.\n    page: Page number.\n    page_size: Page size.\n\nReturns:\n    Page of jobs.",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "experiment_run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experiment Run Id"
            }
          },
          {
            "in": "query",
            "name": "original_session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Original Session Id"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "standalone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Standalone"
            }
          },
          {
            "in": "query",
            "name": "worker_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Worker Id"
            }
          },
          {
            "in": "query",
            "name": "execution_target",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExecutionTarget"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Execution Target"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/claim": {
      "post": {
        "description": "Atomically claim pending jobs within a scope for a worker.\n\nStale claimed or running jobs in scope are requeued or timed out\nfirst, and the claim bumps the worker's last seen time. An unscoped\nclaim yields only pool-target work. With an experiment run id the\nfirst claim moves a pending run to running, and canceling and\nterminal runs yield no jobs.\n\nClients observe HTTP 200 on success, 404 when no worker or scoped\nexperiment run has the referenced id, and 422 on invalid input.\n\nArgs:\n    body: Job claim request.\n    service: Job service.\n    actor: Caller context.\n\nReturns:\n    Claimed jobs.",
        "operationId": "claim_jobs_v1_jobs_claim_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobClaimResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "delete": {
        "description": "Delete a standalone job.\n\nDeletes the job's config when nothing else references it.\n\nClients observe HTTP 204 on success, 404 when no job has this id,\nand 409 when the job belongs to an experiment run or is claimed or\nrunning.\n\nArgs:\n    job_id: Id of the job.\n    service: Job service.\n    actor: Caller context.",
        "operationId": "delete_job_v1_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "Get a job by id.\n\nClients observe HTTP 200 on success and 404 when no job has this\nid.\n\nArgs:\n    job_id: Id of the job.\n    service: Job service.\n    actor: Caller context.\n\nReturns:\n    Stored job.",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      },
      "patch": {
        "description": "Transition a job through the runner status updates.\n\nCompleting stores the scoring result and the computed diff summary.\nThe transition that makes the last job of a run terminal also\nfinalizes the run.\n\nClients observe HTTP 200 on success, 404 when no job has this id,\n409 when the transition is illegal or completing without a linked\nresult session, and 422 when completing without a scoring result or\nfailing without an error.\n\nArgs:\n    job_id: Id of the job.\n    body: Job update request.\n    service: Job service.\n    actor: Caller context.\n\nReturns:\n    Updated job.",
        "operationId": "update_job_v1_jobs__job_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/claim": {
      "post": {
        "description": "Claim a standalone job for a worker.\n\nA stale claim or start is requeued or timed out first, so a job\nwhose worker died is claimable again.\n\nClients observe HTTP 200 on success, 404 when no job has this id,\n409 when the job belongs to an experiment run or is not pending\nafter the staleness resolution, and 422 on invalid input.\n\nArgs:\n    job_id: Id of the job.\n    body: Standalone job claim request.\n    service: Job service.\n    actor: Caller context.\n\nReturns:\n    Claimed job.",
        "operationId": "claim_job_v1_jobs__job_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandaloneJobClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/diff": {
      "get": {
        "description": "Compute the full diff between a job's sessions.\n\nClients observe HTTP 200 on success, 404 when no job has this id,\nand 409 when the job has no result session yet.\n\nArgs:\n    job_id: Id of the job.\n    service: Job service.\n    actor: Caller context.\n\nReturns:\n    Computed replay diff.",
        "operationId": "get_job_diff_v1_jobs__job_id__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayDiffResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Diff",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/heartbeat": {
      "post": {
        "description": "Record a worker heartbeat on a job.\n\nTerminal jobs record nothing and report the stop flag, so the\nworker abandons the job.\n\nClients observe HTTP 200 on success, including terminal jobs, 404\nwhen no job has this id, and 409 when the job is pending.\n\nArgs:\n    job_id: Id of the job.\n    service: Job service.\n    actor: Caller context.\n\nReturns:\n    Heartbeat response with the status and stop flag.",
        "operationId": "heartbeat_job_v1_jobs__job_id__heartbeat_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobHeartbeatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heartbeat Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/release": {
      "post": {
        "description": "Requeue a claimed or running job for another attempt.\n\nClients observe HTTP 200 on success, 404 when no job has this id,\nand 409 when the job is not claimed or running.\n\nArgs:\n    job_id: Id of the job.\n    service: Job service.\n    actor: Caller context.\n\nReturns:\n    Requeued job.",
        "operationId": "release_job_v1_jobs__job_id__release_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/retry": {
      "post": {
        "description": "Requeue a finished standalone job for another attempt.\n\nClients observe HTTP 200 on success, 404 when no job has this id,\nand 409 when the job belongs to an experiment run or is not\nfailed, timed out, or canceled.\n\nArgs:\n    job_id: Id of the job.\n    service: Job service.\n    actor: Caller context.\n\nReturns:\n    Requeued job.",
        "operationId": "retry_job_v1_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}